Replace ground Python scraper with Go CDP
Rewrite the opt-in Trainline/SNCF browser fallback to use the repo's existing Go chromedp stack instead of python3/Playwright, remove the Python scraper, and update offline tests around Go seams.

Split oversized helper/test/doc files so tracked files satisfy the canonical <=800 LOC gate without changing package behavior.

Validation: go test ./...; go vet ./...; go run honnef.co/go/tools/cmd/staticcheck@latest ./...; git diff --check; GitNexus detect-changes --scope staged (low risk, 0 affected processes).
