+++ internal/ground/browser_scraper_test.go
@@ -2,274 +2,216 @@ package ground
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"fmt"
 	"os"
-	"os/exec"
 	"strings"
 	"testing"
 	"time"
 )
 
-func requireScraperScriptPath(t *testing.T) string {
-	t.Helper()
-
-	path, err := resolveScraperScriptPath()
-	if errors.Is(err, errScraperScriptNotFound) {
-		t.Skip(err.Error())
-	}
-	if err != nil {
-		t.Fatalf("resolveScraperScriptPath: %v", err)
-	}
-	return path
-}
-
-// TestBrowserScrapeRoutes_Python3NotFound verifies graceful failure when python3
-// is absent. We cannot easily remove python3, so this test only runs when
-// explicitly requested via the TRVL_TEST_NO_PYTHON env var.
-func TestBrowserScrapeRoutes_Python3NotFound(t *testing.T) {
-	if os.Getenv("TRVL_TEST_NO_PYTHON") == "" {
-		t.Skip("set TRVL_TEST_NO_PYTHON=1 to run")
-	}
-	ctx := context.Background()
-	_, err := BrowserScrapeRoutes(ctx, "trainline", "London", "Paris", "2026-04-10", "EUR")
-	if err == nil {
-		t.Error("expected error when python3 unavailable")
-	}
-}
-
-func TestIsExecNotFound(t *testing.T) {
+func TestParseBrowserPrice(t *testing.T) {
 	t.Parallel()
 
 	cases := []struct {
-		name string
-		err  error
-		want bool
+		raw      string
+		fallback string
+		price    float64
+		currency string
+		ok       bool
 	}{
-		{name: "nil", err: nil, want: false},
-		{name: "plain exec not found", err: &exec.Error{Name: "python3", Err: exec.ErrNotFound}, want: true},
-		{name: "wrapped exec not found", err: fmt.Errorf("wrapped: %w", &exec.Error{Name: "python3", Err: exec.ErrNotFound}), want: true},
-		{name: "other error", err: errors.New("boom"), want: false},
+		{raw: "£ 39.50", fallback: "EUR", price: 39.50, currency: "GBP", ok: true},
+		{raw: "€12,30", fallback: "USD", price: 12.30, currency: "EUR", ok: true},
+		{raw: "$99", fallback: "EUR", price: 99, currency: "USD", ok: true},
+		{raw: "", fallback: "EUR", ok: false},
+		{raw: "€ nope", fallback: "EUR", ok: false},
 	}
 
 	for _, tc := range cases {
 		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
+		t.Run(tc.raw, func(t *testing.T) {
 			t.Parallel()
-			if got := isExecNotFound(tc.err); got != tc.want {
-				t.Fatalf("isExecNotFound(%v) = %v, want %v", tc.err, got, tc.want)
+			price, currency, ok := parseBrowserPrice(tc.raw, tc.fallback)
+			if ok != tc.ok {
+				t.Fatalf("ok = %v, want %v", ok, tc.ok)
+			}
+			if !ok {
+				return
+			}
+			if price != tc.price {
+				t.Fatalf("price = %v, want %v", price, tc.price)
+			}
+			if currency != tc.currency {
+				t.Fatalf("currency = %q, want %q", currency, tc.currency)
 			}
 		})
 	}
 }
 
-func TestBrowserScrapeRoutes_MissingScraperScript(t *testing.T) {
-	t.Setenv("TRVL_SCRAPER_PATH", "__definitely_missing_scraper__.py")
+func TestParseTrainlineBrowserText(t *testing.T) {
+	t.Parallel()
 
-	ctx := context.Background()
-	_, err := BrowserScrapeRoutes(ctx, "trainline", "London", "Paris", "2026-04-10", "EUR")
-	if !errors.Is(err, errScraperScriptNotFound) {
-		t.Fatalf("expected missing scraper error, got %v", err)
+	text := `
+	06:31 09:47 £ 79.90
+	08:01 11:18 £ 39.00
+	09:31 12:47 £ 49.00
+	`
+	routes := parseTrainlineBrowserText(text, "London", "Paris", "2026-04-10", "EUR")
+	if len(routes) != 3 {
+		t.Fatalf("routes = %d, want 3", len(routes))
+	}
+	for _, route := range routes {
+		if route.Provider != "trainline" {
+			t.Fatalf("provider = %q, want trainline", route.Provider)
+		}
+		if route.Price != 39 {
+			t.Fatalf("price = %v, want 39", route.Price)
+		}
+		if route.Currency != "GBP" {
+			t.Fatalf("currency = %q, want GBP", route.Currency)
+		}
+		if route.Departure.City != "London" || route.Arrival.City != "Paris" {
+			t.Fatalf("cities not populated: %+v", route)
+		}
+		if !strings.Contains(route.BookingURL, "/london/paris/2026-04-10") {
+			t.Fatalf("booking URL = %q", route.BookingURL)
+		}
 	}
 }
 
-// TestBrowserScrapeRoutes_ScraperScriptExists ensures the scraper.py file is
-// present relative to browser_scraper.go at build time.
-func TestBrowserScrapeRoutes_ScraperScriptExists(t *testing.T) {
-	path := requireScraperScriptPath(t)
-	if path == "" {
-		t.Fatal("scraperScriptPath returned empty string")
+func TestParseTrainlineBrowserText_DailyMinimum(t *testing.T) {
+	t.Parallel()
+
+	routes := parseTrainlineBrowserText("Lowest fare today € 12.30", "Lyon", "Paris", "2026-04-10", "EUR")
+	if len(routes) != 1 {
+		t.Fatalf("routes = %d, want 1", len(routes))
 	}
-	if _, err := os.Stat(path); err != nil {
-		t.Fatalf("scraper.py not readable at %s: %v", path, err)
+	if routes[0].Departure.Time != "2026-04-10" || routes[0].Arrival.Time != "2026-04-10" {
+		t.Fatalf("daily-minimum times not preserved: %+v", routes[0])
+	}
+	if routes[0].Price != 12.30 || routes[0].Currency != "EUR" {
+		t.Fatalf("price = %v %s, want 12.30 EUR", routes[0].Price, routes[0].Currency)
 	}
 }
 
-// TestBrowserScrapeRoutes_ScraperOutputParsing verifies that scraperOutput
-// JSON round-trips correctly into GroundRoute values.
-func TestBrowserScrapeRoutes_ScraperOutputParsing(t *testing.T) {
-	raw := `{
-		"routes": [
-			{
-				"price": 39.00,
-				"currency": "GBP",
-				"departure": "2026-04-10T06:31:00",
-				"arrival": "2026-04-10T09:47:00",
-				"duration": 196,
-				"type": "train",
-				"provider": "eurostar",
-				"transfers": 0,
-				"booking_url": "https://www.thetrainline.com/book/trains/london/paris/2026-04-10"
-			},
-			{
-				"price": 0,
-				"currency": "EUR",
-				"departure": "2026-04-10T07:00:00",
-				"arrival": "2026-04-10T10:00:00",
-				"duration": 180,
-				"type": "train",
-				"provider": "thalys",
-				"transfers": 0
-			}
-		]
-	}`
+func TestBrowserScrapeRoutes_UnsupportedProvider(t *testing.T) {
+	t.Parallel()
 
-	var result scraperOutput
-	if err := json.Unmarshal([]byte(raw), &result); err != nil {
-		t.Fatalf("unmarshal: %v", err)
-	}
-	if len(result.Routes) != 2 {
-		t.Fatalf("expected 2 routes, got %d", len(result.Routes))
-	}
-
-	// Verify first route.
-	r0 := result.Routes[0]
-	if r0.Price != 39.00 {
-		t.Errorf("price = %f, want 39.00", r0.Price)
-	}
-	if r0.Currency != "GBP" {
-		t.Errorf("currency = %q, want GBP", r0.Currency)
-	}
-	if r0.Duration != 196 {
-		t.Errorf("duration = %d, want 196", r0.Duration)
-	}
-	if r0.Provider != "eurostar" {
-		t.Errorf("provider = %q, want eurostar", r0.Provider)
-	}
-	if r0.Transfers != 0 {
-		t.Errorf("transfers = %d, want 0", r0.Transfers)
-	}
-	if r0.BookingURL == "" {
-		t.Error("booking_url should not be empty")
-	}
-
-	// Zero-price route.
-	r1 := result.Routes[1]
-	if r1.Price != 0 {
-		t.Errorf("second route price = %f, want 0", r1.Price)
+	_, err := BrowserScrapeRoutes(context.Background(), "nosuchprovider", "London", "Paris", "2026-04-10", "EUR")
+	if !errors.Is(err, errBrowserScraperUnsupportedProvider) {
+		t.Fatalf("err = %v, want unsupported-provider error", err)
 	}
 }
 
-// TestBrowserScrapeRoutes_ErrorResponse verifies that a scraper error field
-// is surfaced as a Go error.
-func TestBrowserScrapeRoutes_ErrorResponse(t *testing.T) {
-	// We feed a pre-built JSON error response to the parser directly.
-	raw := `{"routes":[],"error":"playwright not installed"}`
-	var result scraperOutput
-	if err := json.Unmarshal([]byte(raw), &result); err != nil {
-		t.Fatalf("unmarshal: %v", err)
-	}
-	if result.Error == "" {
-		t.Error("expected error field to be set")
-	}
-	if result.Error != "playwright not installed" {
-		t.Errorf("error = %q, want 'playwright not installed'", result.Error)
-	}
-}
+func TestBrowserScrapeRoutes_TrainlineUsesGoBrowserSeam(t *testing.T) {
+	orig := browserScraperNavigateText
+	t.Cleanup(func() { browserScraperNavigateText = orig })
 
-// TestScraperPy_SyntaxCheck verifies scraper.py is valid Python 3 syntax.
-// Uses python3 -m py_compile which exits 0 on success, non-zero on error.
-func TestScraperPy_SyntaxCheck(t *testing.T) {
-	path := requireScraperScriptPath(t)
+	var gotURL string
+	browserScraperNavigateText = func(_ context.Context, targetURL string, dwell time.Duration) (string, error) {
+		gotURL = targetURL
+		if dwell != 15*time.Second {
+			t.Fatalf("dwell = %v, want 15s", dwell)
+		}
+		return "06:31 09:47 £39", nil
+	}
 
-	cmd := exec.Command("python3", "-m", "py_compile", path)
-	out, err := cmd.CombinedOutput()
+	routes, err := BrowserScrapeRoutes(context.Background(), "trainline", "London", "Paris", "2026-04-10", "EUR")
 	if err != nil {
-		t.Fatalf("scraper.py syntax error: %v\n%s", err, out)
+		t.Fatalf("BrowserScrapeRoutes: %v", err)
+	}
+	if len(routes) != 1 {
+		t.Fatalf("routes = %d, want 1", len(routes))
+	}
+	if !strings.Contains(gotURL, "thetrainline.com/book/results") {
+		t.Fatalf("target URL = %q", gotURL)
 	}
 }
 
-// TestScraperPy_MissingInputHandling verifies that scraper.py writes a valid
-// JSON error response when given no stdin input (instead of crashing).
-func TestScraperPy_MissingInputHandling(t *testing.T) {
-	path := requireScraperScriptPath(t)
+func TestBrowserScrapeRoutes_SNCFUsesGoBrowserSeam(t *testing.T) {
+	orig := browserScraperSNCFResponses
+	t.Cleanup(func() { browserScraperSNCFResponses = orig })
 
-	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
+	browserScraperSNCFResponses = func(_ context.Context, bookingURL string, fromStation, toStation SNCFStation, date string) ([]map[string]any, string, error) {
+		if !strings.Contains(bookingURL, "/FRPAR/FRLYS/2026-04-10") {
+			t.Fatalf("booking URL = %q", bookingURL)
+		}
+		if fromStation.Code != "FRPAR" || toStation.Code != "FRLYS" || date != "2026-04-10" {
+			t.Fatalf("unexpected SNCF arguments: %+v %+v %s", fromStation, toStation, date)
+		}
+		return []map[string]any{{
+			"journeys": []any{map[string]any{
+				"price":             map[string]any{"amount": 2900.0, "currency": "EUR"},
+				"departureDate":     "2026-04-10T07:00:00",
+				"arrivalDate":       "2026-04-10T09:00:00",
+				"durationInMinutes": 120.0,
+			}},
+		}}, "key", nil
+	}
 
-	cmd := exec.CommandContext(ctx, "python3", path)
-	cmd.Stdin = strings.NewReader("") // empty stdin
-
-	out, err := cmd.Output()
+	routes, err := BrowserScrapeRoutes(context.Background(), "sncf", "Paris", "Lyon", "2026-04-10", "EUR")
 	if err != nil {
-		t.Fatalf("scraper.py crashed on empty input: %v", err)
+		t.Fatalf("BrowserScrapeRoutes: %v", err)
 	}
-
-	var result scraperOutput
-	if err := json.Unmarshal(out, &result); err != nil {
-		t.Fatalf("scraper.py did not output valid JSON: %v\nraw: %s", err, out)
+	if len(routes) != 1 {
+		t.Fatalf("routes = %d, want 1", len(routes))
 	}
-	if result.Error == "" {
-		t.Error("expected error field for empty input")
+	route := routes[0]
+	if route.Provider != "sncf" || route.Price != 2900 {
+		t.Fatalf("route parse mismatch: %+v", route)
 	}
-}
-
-// TestScraperPy_UnknownProvider verifies that scraper.py gracefully handles
-// an unrecognised provider name.
-func TestScraperPy_UnknownProvider(t *testing.T) {
-	path := requireScraperScriptPath(t)
-
-	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-
-	inp, _ := json.Marshal(map[string]string{
-		"provider": "nosuchprovider",
-		"from":     "London",
-		"to":       "Paris",
-		"date":     "2026-04-10",
-		"currency": "EUR",
-	})
-
-	cmd := exec.CommandContext(ctx, "python3", path)
-	cmd.Stdin = strings.NewReader(string(inp))
-
-	out, err := cmd.Output()
-	if err != nil {
-		t.Fatalf("scraper.py crashed on unknown provider: %v", err)
+	if route.Departure.City != "Paris" || route.Departure.Station == "" {
+		t.Fatalf("departure not enriched: %+v", route.Departure)
 	}
-
-	var result scraperOutput
-	if err := json.Unmarshal(out, &result); err != nil {
-		t.Fatalf("scraper.py did not output valid JSON: %v\nraw: %s", err, out)
-	}
-	if result.Error == "" {
-		t.Error("expected error field for unknown provider")
+	if route.Arrival.City != "Lyon" || route.Arrival.Station == "" {
+		t.Fatalf("arrival not enriched: %+v", route.Arrival)
 	}
 }
 
-// TestScraperPy_InvalidJSON verifies that scraper.py handles bad JSON input
-// without panicking and returns an error response.
-func TestScraperPy_InvalidJSON(t *testing.T) {
-	path := requireScraperScriptPath(t)
+func TestCaptureSNCFKey_UsesGoBrowserSeam(t *testing.T) {
+	orig := browserScraperCaptureHeader
+	t.Cleanup(func() { browserScraperCaptureHeader = orig })
 
-	ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-	defer cancel()
-
-	cmd := exec.CommandContext(ctx, "python3", path)
-	cmd.Stdin = strings.NewReader("not valid json {{{")
-
-	out, err := cmd.Output()
-	if err != nil {
-		t.Fatalf("scraper.py crashed on invalid JSON: %v", err)
+	browserScraperCaptureHeader = func(_ context.Context, targetURLs []string, headerName string, dwell time.Duration) (string, error) {
+		if len(targetURLs) != 2 {
+			t.Fatalf("target URLs = %d, want 2", len(targetURLs))
+		}
+		if headerName != "x-bff-key" {
+			t.Fatalf("header = %q, want x-bff-key", headerName)
+		}
+		if dwell != 5*time.Second {
+			t.Fatalf("dwell = %v, want 5s", dwell)
+		}
+		return "live-key", nil
 	}
 
-	var result scraperOutput
-	if err := json.Unmarshal(out, &result); err != nil {
-		t.Fatalf("scraper.py did not output valid JSON: %v\nraw: %s", err, out)
-	}
-	if result.Error == "" {
-		t.Error("expected error field for invalid JSON input")
+	if got := captureSNCFKey(context.Background()); got != "live-key" {
+		t.Fatalf("captureSNCFKey = %q, want live-key", got)
 	}
 }
 
-// TestBrowserScrapeRoutes_Integration runs a live browser scrape for Trainline.
-// Skipped in short mode and when TRVL_TEST_BROWSER is not set (expensive/flaky).
+func TestDecodeBrowserJSONBodies(t *testing.T) {
+	t.Parallel()
+
+	bodies := []string{
+		`{"journeys":[{"price":42}]}`,
+		`{"_httpError":403,"_body":"blocked"}`,
+		`<html>not json</html>`,
+		``,
+	}
+	got := decodeBrowserJSONBodies(bodies)
+	if len(got) != 1 {
+		t.Fatalf("decoded = %d, want 1", len(got))
+	}
+	if _, ok := got[0]["journeys"]; !ok {
+		t.Fatalf("decoded body missing journeys: %+v", got[0])
+	}
+}
+
 func TestBrowserScrapeRoutes_Integration(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping browser integration test in short mode")
 	}
-	if os.Getenv("TRVL_TEST_BROWSER") == "" {
+	if testingEnv := strings.TrimSpace(os.Getenv("TRVL_TEST_BROWSER")); testingEnv == "" {
 		t.Skip("set TRVL_TEST_BROWSER=1 to run browser integration tests")
 	}
 
@@ -277,29 +219,14 @@ func TestBrowserScrapeRoutes_Integration(t *testing.T) {
 	defer cancel()
 
 	date := time.Now().AddDate(0, 1, 0).Format("2006-01-02")
-
 	routes, err := BrowserScrapeRoutes(ctx, "trainline", "London", "Paris", date, "EUR")
 	if err != nil {
 		t.Skipf("browser scraper unavailable: %v", err)
 	}
 	if len(routes) == 0 {
-		t.Skip("no routes returned (page may have changed)")
+		t.Skip("no routes returned")
 	}
-
-	r := routes[0]
-	if r.Price <= 0 {
-		t.Errorf("price = %f, want > 0", r.Price)
-	}
-	if r.Type != "train" && r.Type != "mixed" {
-		t.Errorf("type = %q, want train or mixed", r.Type)
-	}
-	if r.Departure.City == "" {
-		t.Error("departure city should not be empty")
-	}
-	if r.Arrival.City == "" {
-		t.Error("arrival city should not be empty")
-	}
-	if r.BookingURL == "" {
-		t.Error("booking URL should not be empty")
+	if routes[0].Price <= 0 {
+		t.Fatalf("price = %f, want > 0", routes[0].Price)
 	}
 }
